+++ src/AImove.js
@@ -1,4 +1,4 @@
-import { items } from './app2.js';
+import { items } from '../app2.js';
 import { setO } from './setX-setO.js';
 
 function AImove() {
