+++ src/checkWin.js
@@ -1,4 +1,4 @@
-import { it1, it2, it3, it4, it5, it6, it7, it8, it9 } from './app2.js';
+import { it1, it2, it3, it4, it5, it6, it7, it8, it9 } from '../app2.js';
 
 function checkWin() {
   const groups = [
