+++ src/setSymbol.js
